refact more core functions
