+++ src/canva/renderer/Renderer.tsx
@@ -32,6 +32,7 @@ import {
 } from "./core/scrollNodeHandler";
 import { zoomAndPanHandler } from "./core/zoomAndPanHanlder";
 import { scrollContentHorizontallyCenter } from "./core/scrollContentCenter";
+import { fitContentToViewHandler } from "./core/fitContentToViewHandler";
 
 interface ReactInfiniteCanvasRendererProps extends ReactInfiniteCanvasProps {
   children: any;
@@ -71,6 +72,31 @@ export const ReactInfiniteCanvasRenderer = memo(
       translateY: 0,
       scale: 1,
     });
+
+    const fitContentToView = useCallback(
+      (args) =>
+        fitContentToViewHandler({
+          ...args,
+          flowRendererRef,
+          canvasRef,
+          d3Selection,
+          setZoomTransform,
+          scrollBarRef,
+          d3Zoom,
+          minZoom,
+          maxZoom,
+        }),
+      [
+        canvasRef,
+        d3Selection,
+        d3Zoom,
+        flowRendererRef,
+        maxZoom,
+        minZoom,
+        scrollBarRef,
+        setZoomTransform,
+      ]
+    );
 
     useImperativeHandle(ref, () => ({
       scrollNodeToCenter: scrollNodeToCenterHandler,
@@ -158,65 +184,6 @@ export const ReactInfiniteCanvasRenderer = memo(
         -(scrollDelta.deltaY / currentZoom)
       );
     };
-
-    const fitContentToView = useCallback(
-      function fitContentHandler({
-        duration = 500,
-        offset = { x: 0, y: 0 },
-        scale,
-        maxZoomLimit = ZOOM_CONFIGS.FIT_TO_VIEW_MAX_ZOOM,
-      }: {
-        duration?: number;
-        offset?: { x: number; y: number };
-        scale?: number;
-        maxZoomLimit?: number;
-      }) {
-        requestIdleCallback(
-          () => {
-            if (!flowRendererRef.current) return;
-            const canvasNode = select(canvasRef.current);
-            const contentBounds =
-              flowRendererRef.current.getBoundingClientRect();
-            const currentZoom = d3Selection.current.property("__zoom").k || 1;
-            const containerBounds = canvasRef.current?.getBoundingClientRect();
-            const { width: containerWidth = 0, height: containerHeight = 0 } =
-              containerBounds || {};
-            const scaleDiff = 1 / currentZoom;
-            const contentWidth = contentBounds.width * scaleDiff;
-            const contentHeight = contentBounds.height * scaleDiff;
-            const heightRatio = containerHeight / contentHeight;
-            const newScale =
-              scale ??
-              clampValue({
-                value: Math.min(maxZoomLimit, heightRatio),
-                min: minZoom,
-                max: maxZoom,
-              });
-
-            // below code calculates the translateX and translateY values to
-            // center the content horizontally and fit content vertically
-            const translateX =
-              (containerWidth - contentWidth * newScale) / 2 + offset.x;
-            const translateY = offset.y;
-
-            const newTransform = zoomIdentity
-              .translate(translateX, translateY)
-              .scale(newScale);
-            setZoomTransform({ translateX, translateY, scale: newScale });
-            scrollBarRef.current?.resetScrollPos();
-
-            canvasNode
-              // @ts-ignore
-              .transition()
-              .duration(duration)
-              .call(d3Zoom.transform, newTransform);
-          },
-          { timeout: TIME_TO_WAIT }
-        );
-      },
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-      [maxZoom, minZoom]
-    );
 
     const getCanvasState = () => {
       return {
